scoring tracker moved
